Can now make a fresh copy of an existing Workflow.
There are occasions in which a new Workflow needs to be created from an
existing one, but with updated Firework ids to ensure no collisions with
any Workflows it is appended to. Such collisions can occur when a
Workflow has been serialized and stored within a FireWork spec, then
used by that FireWork to add additional Fireworks to the Workflow
itself. The result of these collisions are often malformed links.

Building a fresh Workflow from that stored avoids this problem, since
the member Fireworks are given fresh ids that have not been used within
the current session.