Feature request - adding state of the FW for _pass_job_info
For the special spec argument "_pass_job_info", it would be useful to get the state of the parent FWs. Normally, the state of a parent should be "COMPLETED", however, if we also have the special spec argument "_allow_fizzled_parents" this is not always the case. It would be easier to pass the job info for this situation than to load it from the collection during a FW.