Quick fix for #416
`webbrowser.open()` doesn't seem to need additional parallelization, but spawn it in a thread just so it can delay to give the web server more time to get ready to serve pages. In any case, not spawning a webserver process fixes the Issue.

When `webbrowser.open(url)` has to launch Chrome, Chrome opens the Fireworks `url` in a new window then restores its previous tabs in another window right on top of the Fireworks window, kind of hiding it.